# app/ai-service/api/routes.py
import io
import time
from typing import Annotated, Optional

from fastapi import APIRouter, File, Form, HTTPException, Request, UploadFile
import metrics
from slowapi import Limiter
from slowapi.util import get_remote_address

from schemas.ocr import OCRData, OCRFieldResult, OCRResponse, LanguageHint
from services.ocr import OCRService
from services.ocr_confidence import assess_confidence
from config import settings

router = APIRouter(tags=["ai"])
limiter = Limiter(key_func=get_remote_address)

ALLOWED_CONTENT_TYPES = {
    "image/jpeg",
    "image/png",
    "image/jpg",
    "image/bmp",
    "image/tiff",
    "image/webp",
}

ocr_service = OCRService()


@router.post("/ai/ocr")
@limiter.limit(settings.request_rate_limit)
async def process_ocr(
    request: Request,
    image: Annotated[UploadFile, File(description="Image file to process")],
    language_hint: Annotated[
        Optional[LanguageHint], Form(description="Language hint for OCR")
    ] = None,
    document_type: Annotated[
        Optional[str],
        Form(
            description=(
                "Optional document type (e.g. id_card, passport) used to pick "
                "a per-document-type review threshold"
            )
        ),
    ] = None,
) -> OCRResponse:
    start_time = time.time()

    if image.content_type not in ALLOWED_CONTENT_TYPES:
        raise HTTPException(
            status_code=400,
            detail={
                "code": "invalid_content_type",
                "message": f"Invalid content type: {image.content_type}. Allowed: {', '.join(ALLOWED_CONTENT_TYPES)}",
            },
        )

    try:
        contents = await image.read()

        if len(contents) == 0:
            raise HTTPException(
                status_code=400,
                detail={
                    "code": "empty_image",
                    "message": "Uploaded image is empty",
                },
            )

        from PIL import Image

        try:
            img = Image.open(io.BytesIO(contents))
        except Exception as e:
            raise HTTPException(
                status_code=400,
                detail={
                    "code": "invalid_image",
                    "message": f"Could not decode image: {str(e)}",
                },
            )

        start_inference = time.time()
        result = ocr_service.process_image(
            img, language_hint=language_hint.value if language_hint else None
        )
        inference_latency = time.time() - start_inference

        metrics.INFERENCE_LATENCY.labels(task_type="ocr").observe(inference_latency)
        metrics.logger.info(f"OCR Inference completed in {inference_latency:.4f}s")

        processing_time_ms = int((time.time() - start_time) * 1000)

        # Band the extraction so a low-confidence document is flagged for a
        # human instead of being treated as authoritative (issue #984).
        assessment = assess_confidence(
            [field.confidence for field in result.fields.values()],
            document_type=document_type,
        )

        return OCRResponse(
            success=True,
            data=OCRData(
                fields={
                    name: OCRFieldResult(value=field.value, confidence=field.confidence)
                    for name, field in result.fields.items()
                },
                raw_text=result.raw_text,
                processing_time_ms=processing_time_ms,
                confidence=assessment.confidence,
                confidence_band=assessment.band,
                needs_review=assessment.needs_review,
                review_threshold=assessment.review_threshold,
                document_type=document_type,
            ),
            processing_time_ms=processing_time_ms,
        )

    except HTTPException:
        raise
    except Exception as e:
        processing_time_ms = int((time.time() - start_time) * 1000)
        return OCRResponse(
            success=False,
            error={
                "code": "processing_error",
                "message": str(e),
            },
            processing_time_ms=processing_time_ms,
        )