# app/ai-service/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_process_ocr_ai_ocr_post": {
        "properties": {
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional document type (e.g. id_card, passport) used to pick a per-document-type review threshold",
            "title": "Document Type"
          },
          "image": {
            "description": "Image file to process",
            "format": "binary",
            "title": "Image",
            "type": "string"
          },
          "language_hint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguageHint"
              },
              {
                "type": "null"
              }
            ],
            "description": "Language hint for OCR"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_process_ocr_ai_ocr_post",
        "type": "object"
      },
      "Body_process_ocr_v1_ai_ocr_post": {
        "properties": {
          "anchor_metadata": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JSON encoded AnchorMetadata",
            "title": "Anchor Metadata"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional document type (e.g. id_card, passport) used to pick a per-document-type review threshold",
            "title": "Document Type"
          },
          "image": {
            "description": "Image file to process",
            "format": "binary",
            "title": "Image",
            "type": "string"
          },
          "language_hint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguageHint"
              },
              {
                "type": "null"
              }
            ],
            "description": "Language hint for OCR"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_process_ocr_v1_ai_ocr_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InferenceRequest": {
        "description": "Request model for AI inference endpoints.",
        "example": {
          "campaign_id": "123e4567-e89b-12d3-a456-426614174000",
          "claim_id": "123e4567-e89b-12d3-a456-426614174001",
          "data": {
            "document_url": "https://example.com/doc.pdf"
          },
          "network": "testnet",
          "package_id": "pkg_abc123def",
          "type": "verification"
        },
        "properties": {
          "campaign_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "claim_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Id"
          },
          "contract_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Address"
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "network": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "testnet",
            "title": "Network"
          },
          "package_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Id"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "normal",
            "title": "Priority"
          },
          "transaction_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Hash"
          },
          "type": {
            "default": "inference",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "InferenceRequest",
        "type": "object"
      },
      "LanguageHint": {
        "enum": [
          "eng",
          "spa",
          "fra",
          "deu",
          "ita",
          "por",
          "chi_sim",
          "ara",
          "hin",
          "jpn"
        ],
        "title": "LanguageHint",
        "type": "string"
      },
...
      "OCRResponse": {
        "examples": [
          {
            "anchor_metadata": {
              "campaign_ref": "campaign-2024-001",
              "claim_id": "claim-abc123"
            },
            "data": {
              "fields": {
                "full_name": {
                  "confidence": 0.95,
                  "value": "John Doe"
                },
                "id_number": {
                  "confidence": 0.9,
                  "value": "123456789"
                }
              },
              "processing_time_ms": 1500,
              "raw_text": "John Doe\nID: 123456789"
            },
            "processing_time_ms": 1500,
            "success": true
          },
          {
            "error": {
              "code": "invalid_image",
              "message": "Could not decode image"
            },
            "processing_time_ms": 500,
            "success": false
          }
        ],
        "properties": {
          "anchor_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnchorMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OCRData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              {
                "code": "invalid_image",
                "message": "Could not decode image"
              }
            ],
            "title": "Error"
          },
          "processing_time_ms": {
            "examples": [
              1500
            ],
            "title": "Processing Time Ms",
            "type": "integer"
          },
          "success": {
            "examples": [
              true
            ],
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "processing_time_ms"
        ],
        "title": "OCRResponse",
        "type": "object"
      },
...
      "ResultEnvelope_OCRData_": {
        "properties": {
          "anchor_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnchorMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Aggregate confidence score in [0, 1].",
            "examples": [
              0.92
            ],
            "title": "Confidence"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Version of the prompt template used for verification.",
            "examples": [
              "v1"
            ],
            "title": "Prompt Version"
          },
          "reasons": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable explanations or reasons for the result.",
            "examples": [
              [
                "Liveness verification passed"
              ]
            ],
            "title": "Reasons"
          },
          "result": {
            "$ref": "#/components/schemas/OCRData"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Request-scoped correlation ID for distributed tracing.",
            "examples": [
              "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "result"
        ],
        "title": "ResultEnvelope[OCRData]",
        "type": "object"
      },
...
      "TaskStatusResponse": {
        "description": "Response model for task status.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI service layer for Soter platform using FastAPI",
    "title": "Soter AI Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ai/ocr": {
      "post": {
        "operationId": "process_ocr_ai_ocr_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_ocr_ai_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Ocr",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/v1/ai/inference": {
      "post": {
        "description": "Create a background task for heavy AI inference.\n\nOffloads time-consuming AI tasks to background workers. Use the\nreturned ``task_id`` to poll for results via ``GET /ai/status/{task_id}``.\n\nContract-aware metadata fields are used to anchor verification results\nto on-chain events during Testnet demos.",
        "operationId": "create_inference_task_v1_ai_inference_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Inference Task",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/ai/jobs/{task_id}": {
      "get": {
        "description": "Get the current status of a queued AI job.\n\nThis is the canonical poll endpoint for backend clients. Possible\nstatus values: ``pending``, ``processing``, ``retrying``, ``completed``,\n``failed``, ``cancelled``, ``expired``, ``timed_out``. A job that never\nreceived a worker within the configured idle window reports\n``timed_out`` with an explanatory ``error``.",
        "operationId": "get_job_status_v1_ai_jobs__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/ai/ocr": {
      "post": {
        "description": "Extract text fields from an uploaded document image.",
        "operationId": "process_ocr_v1_ai_ocr_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_ocr_v1_ai_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEnvelope_OCRData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Ocr",
        "tags": [
          "ocr"
        ]
      }
    },
...
    "/v1/ai/status/{task_id}": {
      "get": {
        "description": "Get the current status of a background inference task.\n\nPoll this endpoint after creating a task. Possible status values:\n``pending``, ``processing``, ``completed``, ``failed``, ``cancelled``,\n``expired``, ``timed_out``. A job still queued past the configured\nidle window reports ``timed_out`` rather than ``pending`` forever.\n\nWhen completed, the result will include contract-aware metadata\nthat can be anchored to on-chain events.",
        "operationId": "get_task_status_v1_ai_status__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task Status",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/ai/task/{task_id}/cancel": {
      "post": {
        "description": "Cancel a pending or in-progress inference task.",
        "operationId": "cancel_task_v1_ai_task__task_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Task",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/ai/task/{task_id}/expire": {
      "post": {
        "description": "Expire a pending or in-progress inference task.",
        "operationId": "expire_task_v1_ai_task__task_id__expire_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Expire Task",
        "tags": [
          "inference"
        ]
      }
    },
...
  }
}